Use _only_ hashes for verifying traces
This reverts the work in previous commit that started using values only
for verifying traces, and was prompted by wanting to easily serialise
the traces to disk for later re-use. If we store the values in the
traces we duplicate a lot of values, which blows up the size when
(somewhat naively) serializing the tree to disk. Using only hashes
avoids this.