Fix race condition in memoiseWithCycleDetection
Keep track of the threadids that are waiting for the result of a query
and make sure that they are removed from the deps map _before_ the
result of the query is written to the result MVar.

As it was before the threadids were sometimes removed too late from the
deps map, which could mean false positive `Cyclic` exceptions.